refactor: move instruction composition to separate files
